Feature:add sw trigger on MDSplus event
