+++ java/jdevices/src/main/resources/RFX_RPADCSetup.form
@@ -21,7 +21,7 @@
     <AuxValue name="FormSettings_listenerGenerationStyle" type="java.lang.Integer" value="0"/>
     <AuxValue name="FormSettings_variablesLocal" type="java.lang.Boolean" value="false"/>
     <AuxValue name="FormSettings_variablesModifier" type="java.lang.Integer" value="2"/>
-    <AuxValue name="designerSize" type="java.awt.Dimension" value="-84,-19,0,5,115,114,0,18,106,97,118,97,46,97,119,116,46,68,105,109,101,110,115,105,111,110,65,-114,-39,-41,-84,95,68,20,2,0,2,73,0,6,104,101,105,103,104,116,73,0,5,119,105,100,116,104,120,112,0,0,1,113,0,0,2,-10"/>
+    <AuxValue name="designerSize" type="java.awt.Dimension" value="-84,-19,0,5,115,114,0,18,106,97,118,97,46,97,119,116,46,68,105,109,101,110,115,105,111,110,65,-114,-39,-41,-84,95,68,20,2,0,2,73,0,6,104,101,105,103,104,116,73,0,5,119,105,100,116,104,120,112,0,0,1,113,0,0,3,-123"/>
   </AuxValues>
 
   <Layout class="org.netbeans.modules.form.compat2.layouts.DesignBorderLayout"/>
@@ -42,7 +42,7 @@
 
       <Layout class="org.netbeans.modules.form.compat2.layouts.DesignGridLayout">
         <Property name="columns" type="int" value="1"/>
-        <Property name="rows" type="int" value="6"/>
+        <Property name="rows" type="int" value="7"/>
       </Layout>
       <SubComponents>
         <Container class="javax.swing.JPanel" name="jPanel2">
@@ -171,7 +171,6 @@
               <Properties>
                 <Property name="identifier" type="java.lang.String" value=""/>
                 <Property name="labelString" type="java.lang.String" value="Pre Samples: "/>
-                <Property name="numCols" type="int" value="4"/>
                 <Property name="offsetNid" type="int" value="12"/>
               </Properties>
             </Component>
@@ -182,18 +181,40 @@
                 <Property name="offsetNid" type="int" value="13"/>
               </Properties>
             </Component>
+            <Component class="DeviceField" name="deviceField11">
+              <Properties>
+                <Property name="identifier" type="java.lang.String" value=""/>
+                <Property name="labelString" type="java.lang.String" value="MDSplus trig. event"/>
+                <Property name="offsetNid" type="int" value="28"/>
+                <Property name="textOnly" type="boolean" value="true"/>
+              </Properties>
+            </Component>
+          </SubComponents>
+        </Container>
+        <Container class="javax.swing.JPanel" name="jPanel8">
+
+          <Layout class="org.netbeans.modules.form.compat2.layouts.DesignFlowLayout"/>
+          <SubComponents>
             <Component class="DeviceField" name="deviceField10">
               <Properties>
                 <Property name="identifier" type="java.lang.String" value=""/>
-                <Property name="labelString" type="java.lang.String" value="Event Level:"/>
+                <Property name="labelString" type="java.lang.String" value="Ev. streaming level:"/>
                 <Property name="numCols" type="int" value="4"/>
                 <Property name="offsetNid" type="int" value="10"/>
+              </Properties>
+            </Component>
+            <Component class="DeviceField" name="deviceField6">
+              <Properties>
+                <Property name="identifier" type="java.lang.String" value=""/>
+                <Property name="labelString" type="java.lang.String" value="Ev. streaming samples: "/>
+                <Property name="numCols" type="int" value="4"/>
+                <Property name="offsetNid" type="int" value="11"/>
               </Properties>
             </Component>
             <Component class="DeviceField" name="deviceField3">
               <Properties>
                 <Property name="identifier" type="java.lang.String" value=""/>
-                <Property name="labelString" type="java.lang.String" value="Event code: "/>
+                <Property name="labelString" type="java.lang.String" value="Timing Highway event code: "/>
                 <Property name="numCols" type="int" value="4"/>
                 <Property name="offsetNid" type="int" value="23"/>
               </Properties>
@@ -208,16 +229,8 @@
               <Properties>
                 <Property name="identifier" type="java.lang.String" value=""/>
                 <Property name="labelString" type="java.lang.String" value="Trigger: "/>
-                <Property name="numCols" type="int" value="30"/>
+                <Property name="numCols" type="int" value="35"/>
                 <Property name="offsetNid" type="int" value="5"/>
-              </Properties>
-            </Component>
-            <Component class="DeviceField" name="deviceField6">
-              <Properties>
-                <Property name="identifier" type="java.lang.String" value=""/>
-                <Property name="labelString" type="java.lang.String" value="Event Samples: "/>
-                <Property name="numCols" type="int" value="4"/>
-                <Property name="offsetNid" type="int" value="11"/>
               </Properties>
             </Component>
           </SubComponents>
